When the workbench window is out of the (default) screen, center it.
Before this change, the window x1,y1 would be adapted while x2,y2
would be fixed (kind of). This would sometimes result in minimal
window widths that would trigger a Qt warning on the console saying

QWindowsWindow::setGeometry: Unable to set geometry [...]

I changed the behavior so that the window will be centered in a
default size of 90% when it does not show on the default screen.
This avoids unusable tiny window sizes.

Signed-off-by: Daniel Maleike <code@maleike.de>